Moved "IgnoreBorder" to "World" category
