+++ src/main/java/meteordevelopment/meteorclient/mixin/WorldBorderMixin.java
@@ -6,7 +6,7 @@
 package meteordevelopment.meteorclient.mixin;
 
 import meteordevelopment.meteorclient.systems.modules.Modules;
-import meteordevelopment.meteorclient.systems.modules.player.IgnoreBorder;
+import meteordevelopment.meteorclient.systems.modules.world.IgnoreBorder;
 import net.minecraft.entity.Entity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Box;
